Quit Chromes with Cmd+Q, fix Chromes preferences shortcut
Chromes on Linux refuse to respond to Cmd+Q (physical Alt+Q, logical RC+Q under Kinto remapping), so it needs to be mapped to Alt+F4 (M-F4). 

To get Chrome browsers to respond properly to Cmd+Q (Quit), I had to rearrange the order of the `chromeStr` block so that it can take precedence over the more "general" browser block which defines `K("RC-Q"): K("RC-Q"),` and was therefore interfering. 

The preferences shortcut for Chromes was not working, opens the Edit menu but can't proceed further because the "Preferences" item doesn't seem to have a shortcut letter anymore. (Moving through the Edit menu with arrow keys shows no underlined letter in "Preferences" and there is no shortcut visible.) So I implemented a multi-keystroke macro solution modeled on the existing method for opening Firefox preferences. It opens a new tab and just types "chrome://settings" and Enter key. Seems to work just as reliably as the equivalent Firefox preferences shortcut. Possibly better, because it's a bit shorter (enters quicker). Hard to tell. 

Reworked some of the related comments and code block labels based on the Finder mod comments, for clarification of the purpose of each code block in the "browsers" section.