Shortcut for LibreOffice Find and Replace dialog
Shortcut for the Find & Replace dialog in both LibreOffice and Microsoft Excel in Windows or Linux is Ctrl+H. This conflicts with the remapped Cmd+H (Ctrl+H) shortcut used by macOS for Hide Window. 

Excel on macOS resolves the conflict with Cmd+H by using Cmd+Shift+H as an alternative. 

LibreOffice on macOS resolves the conflict with Cmd+H by using Cmd+Option+F, unlike Excel. (Ctrl+Alt+F on a Kinto remapped PC keyboard.) 

The Ctrl+Alt+F shortcut is useful as a system shortcut to open a file browser window, as a match for the common system shortcut of Ctrl+Alt+T to open a terminal window. I am therefore leaving that line commented out by default here, and choosing instead to enable the shortcut that matches Excel, which more users coming from macOS to Linux will probably be familiar with. 

So the remapping that will be enabled by default in all LibreOffice apps with this change is the equivalent of Cmd+Shift+H (Ctrl+Shift+H) to Ctrl+H to activate the Find & Replace dialog.