Tab navigation in terminals with Cmd+Shift+Braces
This update activates tab navigation (prior/next tab, left/right) in terminals, with Ctrl+Shift+Braces (Cmd+Shift+Braces in macOS). 

Verified this is a working shortcut in macOS Mojave, on a 2011 MacBook Pro. Works for tab nav in both Apple's Terminal.app and iTerm2. The Ctrl+PgUp/PgDn shortcut that works for terminal tab nav in Linux does nothing in macOS. 

This is the same shortcut remap that is working for tab nav in web browsers and file managers. All tested working in macOS.