Staging modmap for toggle-able Media Arrows Fix
This is a more elegant solution to the need to fix some PC laptop or compact keyboards with media functions on Fn+arrow keys to have PgUp/PgDn/Home/End nav keys on the Fn+arrows, like an Apple laptop keyboard. 

Disabled modmap lines, meant to be optionally enabled by the user, manually or with a script or tray menu tweak item (future). All lines tagged with `(media_arrows_fix)`.