[Win] Typo/format cleanup in optspecialchars
No substantive changes. Just fixing some inconsistent formatting and missing special character examples in comments for the Option-key special character entry scheme. All comments in that section should now be normalized to the same pattern, with a reference sample for each supported character. 

```
; Description {UNICODE}: SAMPLECHAR (Alt Code)
```

If `(Alt Code)` is missing it's because no Alt Code was found for that character.