[Linux] Option key special character entry scheme
This is a DRAFT of the Option key special character entry scheme (OptSpecialChars) to mirror the same feature set recently implemented in Windows with AutoHotKey. Probably not in its final form, but fully functional already. 

This will REQUIRE `keyszer` v0.7.0 or later for testing. It uses the new Unicode processing helper function that will be in the next release of `keyszer`. 

Not as advanced as the Windows/AHK version due to current limitations with nested key actions. But all the same characters and "dead keys" accents are supported. 

Starts disabled and can be enabled/disabled with Shift+Opt+Cmd+O (letter "O" for "Option"), just like the Windows version. Shows alert notifications when scheme is enabled or disabled, and when the user tries to use the Apple logo character (Shift+Opt+K), which like in Windows requires the Baskerville Old Face font to display properly. 

Other than the Apple logo issue, no known problems. 

No menu item for Linux is implemented yet.