Add support for non-default XDG user dir
In an install script for the Linux.

According to the <https://wiki.archlinux.org/title/XDG_user_directories>
and <https://www.freedesktop.org/wiki/Software/xdg-user-dirs/>, Linux
users can customize their "user directories" to be in a different paths
than the default `~/Downloads` etc. _However_, the install script
hardcoded this path so far, which caused a failure for such users. This
commit is supposed to fix that.