Make Finder Mods active on desktop, block wordwise shortcut
This will make the Finder Mods shortcuts active on the Windows desktop (which has a different `ahk_class` than the usual Explorer windows). Also, this will block the improper action of the "wordwise" shortcut Cmd+Backspace while on the Windows desktop, unless the user is editing a file/folder name on the desktop. The wordwise shortcut was doing "select all icons to top-left corner, delete". Cmd+Backspace should still work when in a "text edit" kind of area, like when renaming files/folders on the desktop.